+++ .gitignore
@@ -2,6 +2,7 @@
 .vscode
 .env
 *.code-workspace
+.ipynb_checkpoints
 
 /build/
 __pycache__
