+++ lib/Backend/IREE/Passes.cpp
@@ -16,20 +16,21 @@ using namespace mlir::NPCOMP;
 using namespace mlir::NPCOMP::IREEBackend;
 
 namespace {
-// This pass lowers the public ABI of the module to the primitives exposed by
-// the refbackrt dialect.
-class LowerLinkagePass : public LowerLinkageBase<LowerLinkagePass> {
-  void runOnOperation() override {
-    ModuleOp module = getOperation();
-    for (auto func : module.getOps<FuncOp>()) {
-      if (func.getVisibility() == SymbolTable::Visibility::Public)
-        func->setAttr("iree.module.export", UnitAttr::get(&getContext()));
-    }
-  }
-};
-} // namespace
+#define GEN_PASS_REGISTRATION
+#include "npcomp/Backend/IREE/Passes.h.inc"
+} // end namespace
 
-std::unique_ptr<OperationPass<ModuleOp>>
-mlir::NPCOMP::IREEBackend::createLowerLinkagePass() {
-  return std::make_unique<LowerLinkagePass>();
+void mlir::NPCOMP::IREEBackend::createNpcompBackendToIreeFrontendPipeline(
+    OpPassManager &pm) {
+  pm.addPass(createLowerLinkagePass());
+}
+
+void mlir::NPCOMP::IREEBackend::registerIREEBackendPasses() {
+  ::registerPasses();
+
+  mlir::PassPipelineRegistration<>(
+      "npcomp-backend-to-iree-frontend-pipeline",
+      "Pipeline lowering the npcomp backend contract IR to IREE's frontend "
+      "contract.",
+      mlir::NPCOMP::IREEBackend::createNpcompBackendToIreeFrontendPipeline);
 }
